refactor package and set module Global
