+++ lib/gcloud-secret-manager.constant.ts
@@ -1,1 +1,1 @@
-export const GCLOUD_SECRET_MANAGER_LIST = 'GCLOUD_SECRET_MANAGER_LIST';
+export const GCLOUD_SECRET_MANAGER_PARENT = 'GCLOUD_SECRET_MANAGER_PARENT';
